pyproject.toml
[project]
name = "fgd-presentation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "tensorflow>=2.20.0",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "FGD_presentation",
]
